# perf: avoid duplicate account generation