--- a/src/ape/api/accounts.py
+++ b/src/ape/api/accounts.py
@@ -1,5 +1,6 @@
 import os
 from collections.abc import Iterator
+from functools import cached_property
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Optional, Union
 
@@ -443,11 +444,11 @@ class AccountContainerAPI(BaseInterfaceModel):
             Iterator[:class:`~ape.api.accounts.AccountAPI`]
         """
 
-    @property
+    @cached_property
     def data_folder(self) -> Path:
         """
         The path to the account data files.
-        Defaults to ``$HOME/.ape/<plugin_name>`` unless overriden.
+        Defaults to ``$HOME/.ape/<plugin_name>`` unless overridden.
         """
         path = self.config_manager.DATA_FOLDER / self.name
         path.mkdir(parents=True, exist_ok=True)
@@ -573,18 +574,15 @@ class TestAccountContainerAPI(AccountContainerAPI):
     ``AccountContainerAPI`` directly. Then, they show up in the ``accounts`` test fixture.
     """
 
-    @property
+    @cached_property
     def data_folder(self) -> Path:
         """
         **NOTE**: Test account containers do not touch
-        persistant data. By default and unless overriden,
+        persistent data. By default and unless overridden,
         this property returns the path to ``/dev/null`` and
         it is not used for anything.
         """
-        if os.name == "posix":
-            return Path("/dev/null")
-
-        return Path("NUL")
+        return Path("/dev/null" if os.name == "posix" else "NUL")
 
     @abstractmethod
     def generate_account(self) -> "TestAccountAPI":