--- a/src/ape/managers/project.py
+++ b/src/ape/managers/project.py
@@ -1887,7 +1887,7 @@ class Project(ProjectManager):
         self._config_override = overrides
         _ = self.config
 
-        self.account_manager.test_accounts._reset()
+        self.account_manager.test_accounts.reset()
 
     def extract_manifest(self) -> PackageManifest:
         return self.manifest