--- a/src/ape_test/provider.py
+++ b/src/ape_test/provider.py
@@ -351,10 +351,6 @@ class LocalProvider(TestProviderAPI, Web3Provider):
             private_key.to_hex(),
         )
 
-    # def generate_test_account(self) -> "TestAccountAPI":
-    #     # TODO
-    #     return self.evm_backend.add_account()
-
     def get_virtual_machine_error(self, exception: Exception, **kwargs) -> VirtualMachineError:
         if isinstance(exception, ValidationError):
             match = self._CANNOT_AFFORD_GAS_PATTERN.match(str(exception))